Move all source into `src`, add `files` to package.json
